Add files via upload
This website is built as a part of NammaYatri Hackathon. The website built is a prototype and it has potential to be a successful ride booking website. Although, there are many changes that need to be done before it can be launched as a final product. The website has some basic functionalities such as displaying available rides, providing ride details, allowing users to book a ride and display payment options. However, the website needs more features such as user authentication, real-time updates on ride status, booking confirmation and cancellation options, and more. Moreover, the website needs to be optimized for performance and user experience. Despite the limitations, the idea of this website is surely going to be a great success as it addresses the needs of people looking for an easy and convenient way to book auto rickshaw rides. With further development and refinement, this website has the potential to attract a large user base and become a popular platform for booking auto rickshaw rides